Cleanup EntityUi
